Added: - item details UI styles
