+++ client/src/components/item/Comments.js
@@ -21,8 +21,24 @@ import moment from "moment";
 const titleStyle = {
   fontWeight: "bold",
   fontSize: "1.5rem",
-  color: "#324b4b",
+  color: "#e6e6e6",
   marginBottom: "10px",
+};
+
+const buttonStyle2 = {
+  height: "40px",
+  // width: "90px",
+  borderRadius: "15px",
+  backgroundColor: "#1c2536",
+  fontSize: "15px",
+  padding: "4px 12px",
+  marginRight: "10px",
+  color: "#fff",
+  "&:hover": {
+    backgroundColor: "#ff9717",
+    color: "#1c2536",
+    fontWeight:'600'
+  },
 };
 
 const Comments = ({ postId, comments, setItemData, userId }) => {
@@ -34,7 +50,9 @@ const Comments = ({ postId, comments, setItemData, userId }) => {
           display: "flex",
           flexDirection: "column",
           justifyContent: "center",
-          margin: "10px",
+          padding: "10px 10px",
+          borderRadius: "10px",
+          backgroundColor: "#4a5569",
         }}
       >
         <Typography sx={titleStyle}>Comments</Typography>
@@ -44,6 +62,10 @@ const Comments = ({ postId, comments, setItemData, userId }) => {
               <Grid item xs={12}>
                 <List sx={{ width: "100%" }}>
                   <ListItem
+                    sx={{
+                      backgroundColor: "#e6e6e6",
+                      borderRadius: "10px",
+                    }}
                     alignItems="flex-start"
                     secondaryAction={
                       userId === comment?.userId ? (
@@ -79,41 +101,25 @@ const Comments = ({ postId, comments, setItemData, userId }) => {
                         sx={{
                           backgroundColor: "#2E3643",
                           color: "#fff",
+                          width: "45px",
+                          height: "45px",
                         }}
                       />
                     </ListItemAvatar>
                     <ListItemText
                       primary={
                         <React.Fragment>
-                          <Typography
-                            sx={{ display: "inline", paddingRight: "15px" }}
-                            component="span"
-                            variant="body2"
-                            color="text.primary"
-                          >
-                            {comment?.userDetails?.firstName}{" "}
-                            {comment?.userDetails?.lastName}
-                          </Typography>
-
-                          <Typography
-                            sx={{ display: "inline", paddingRight: "15px" }}
-                            component="span"
-                            variant="body2"
-                            color="text.secondary"
-                          >
-                            |
-                          </Typography>
-
-                          <Typography
-                            sx={{ display: "inline" }}
-                            component="span"
-                            variant="body2"
-                            color="text.primary"
-                          >
-                            {moment(comment?.commentDate).format(
-                              "MMMM Do YYYY, h:mm a"
-                            )}
-                          </Typography>
+                          <div className="flex items-end">
+                            <div className="text-lg font-[600]">
+                              {comment?.userDetails?.firstName}{" "}
+                              {comment?.userDetails?.lastName}
+                            </div>
+                            <div className="text-md ml-[12px]">
+                              {moment(comment?.commentDate).format(
+                                "MMMM Do YYYY, h:mm a"
+                              )}
+                            </div>
+                          </div>
                         </React.Fragment>
                       }
                       disablePadding
@@ -140,55 +146,52 @@ const Comments = ({ postId, comments, setItemData, userId }) => {
             <Alert severity="info">No comments yet!</Alert>
           </div>
         )}
-        <Grid container spacing={2}>
-          <Grid item xs={12}>
-            <TextField
-              id="outlined-multiline-static"
-              label="Add Comment"
-              sx={{
-                marginTop: "10px",
-              }}
-              multiline
-              rows={2}
-              placeholder="Your comment goes here..."
-              variant="outlined"
-              fullWidth
-              value={comment}
-              onChange={(e) => {
-                setComment(e.target.value.trim());
-              }}
-            />
-          </Grid>
-          <Grid item xs={12}>
-            <Button
-              onClick={async () => {
-                if (comment.trim() === "")
-                  toast.error("Seriously? An empty comment?");
-                else {
-                  const { status, data } = await addComment(postId, {
-                    comment,
-                  });
-                  if (status !== 200) toast.error("Failed to add comment!");
-                  else {
-                    setItemData(data?.updatedItem);
-                    setComment("");
-                  }
-                }
-              }}
-              variant="contained"
-              sx={{
-                float: "left",
-                backgroundColor: "#1c2536",
-                "&:hover": {
-                  backgroundColor: "#01AD7B",
-                },
-              }}
-            >
-              Add Comment
-            </Button>
-          </Grid>
-        </Grid>
       </Box>
+      <Grid container spacing={2}>
+        <Grid item xs={12}>
+          <TextField
+            id="outlined-multiline-static"
+            label="Add Comment"
+            sx={{
+              marginTop: "10px",
+            }}
+            multiline
+            rows={2}
+            placeholder="Your comment goes here..."
+            variant="outlined"
+            fullWidth
+            value={comment}
+            onChange={(e) => {
+              setComment(e.target.value.trim());
+            }}
+          />
+        </Grid>
+        <Grid item xs={12}>
+          <Button
+            onClick={async () => {
+              if (comment.trim() === "")
+                toast.error("Seriously? An empty comment?");
+              else {
+                const { status, data } = await addComment(postId, {
+                  comment,
+                });
+                if (status !== 200) toast.error("Failed to add comment!");
+                else {
+                  setItemData(data?.updatedItem);
+                  setComment("");
+                }
+              }
+            }}
+            variant="contained"
+            sx={{
+              ...buttonStyle2,
+              float: "left",
+            }}
+          >
+            Add Comment
+          </Button>
+        </Grid>
+      </Grid>
     </>
   );
 };
